Build: Update package.json for compatibility with NPM modules
Closes GH-98

NPM and other package managers like JSPM expect to have a main
attribute in package.json to specify which file should be returned
when someone requires the library. [1]

It also specifies jQuery >= 1.11.0 as a peerDependency [2]. Older
versions of this package on npm are Node-only shims using contextify
from before we took ownership of the jquery package, they should not
be used and they will not "just work" with this plugin.

[1]: https://docs.npmjs.com/files/package.json#main
[2]: https://docs.npmjs.com/files/package.json#peerdependencies